+++ src/main/java/jnt/scimark2/SOR.java
@@ -1,7 +1,7 @@
 package jnt.scimark2;
 
 public class SOR {
-    public static final double num_flops(int M, int N, int num_iterations) {
+    public static final double num_flops(int M, int N, long num_iterations) {
         double Md = M;
         double Nd = N;
         double num_iterD = num_iterations;
@@ -9,7 +9,7 @@ public class SOR {
         return (Md - 1) * (Nd - 1) * num_iterD * 6.0;
     }
 
-    public static final void execute(double omega, double[][] G, int
+    public static final void execute(double omega, double[][] G, long
             num_iterations) {
         int M = G.length;
         int N = G[0].length;
@@ -21,7 +21,7 @@ public class SOR {
         //
         int Mm1 = M - 1;
         int Nm1 = N - 1;
-        for (int p = 0; p < num_iterations; p++) {
+        for (long p = 0; p < num_iterations; p++) {
             for (int i = 1; i < Mm1; i++) {
                 double[] Gi = G[i];
                 double[] Gim1 = G[i - 1];
